[TINKERPOP-2274] Split the TestHelper from gremlin-core/test into two classes - CoreTestHelper in the gremlin-core/main, so that it poublishes common methods, available to be inherited, whereas the only duplicated code is left in core/test TestHelper not published, thus avoiding name conflict with TestHelper from gremlin-test
