+++ gremlin-core/src/test/java/org/apache/tinkerpop/gremlin/TestHelper.java
@@ -16,17 +16,27 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.tinkerpop.gremlin.structure.io.graphson;
+package org.apache.tinkerpop.gremlin;
 
-import org.apache.tinkerpop.gremlin.TestHelper;
-import org.junit.Test;
+import static org.junit.Assert.assertTrue;
+
+import java.lang.reflect.Constructor;
+import java.lang.reflect.Modifier;
 
 /**
  * @author Stephen Mallette (http://stephen.genoprime.com)
  */
-public class GraphSONUtilTest {
-    @Test
-    public void shouldBeUtilityClass() throws Exception {
-        TestHelper.assertIsUtilityClass(GraphSONUtil.class);
+public class TestHelper {
+
+    protected TestHelper() {}
+    
+    public static void assertIsUtilityClass(final Class<?> utilityClass) throws Exception {
+        final Constructor constructor = utilityClass.getDeclaredConstructor();
+
+        assertTrue(Modifier.isFinal(utilityClass.getModifiers()));
+        assertTrue(Modifier.isPrivate(constructor.getModifiers()));
+        constructor.setAccessible(true);
+        constructor.newInstance();
     }
+        
 }
